Move io functions into subway_io.py module; a bug fixed
